Build: Changed friends to mentores & added details
